requirements.txt
Pillow
numba
numpy
tqdm